split code to improve tree shaking
5bdf9bb didn't quite fix the webpack error apparently
