+++ src/server/gamebananaMirror/yamlHandlers/utils/getFileSystemErrorString.ts
@@ -1,0 +1,2 @@
+
+export const getFileSystemErrorString = (yamlName: string) => `Failed to write the ${yamlName} to the file system.`;
